+++ codigo-fonte/cultivo/.gitignore
