+++ codigo-fonte/cultivo/apps/web/pwa-assets.config.ts
