+++ codigo-fonte/cultivo/apps/web/src/routes/index.tsx
@@ -1,0 +1,35 @@
+import { createFileRoute } from "@tanstack/react-router";
+import { useQuery } from "convex/react";
+import { api } from "@my-cultivoapp/backend/convex/_generated/api";
+
+export const Route = createFileRoute("/")({
+	component: HomeComponent,
+});
+
+
+
+function HomeComponent() {
+	const healthCheck = useQuery(api.healthCheck.get);
+
+	return (
+		<div className="container mx-auto max-w-3xl px-4 py-2">
+			<div className="grid gap-6">
+				<section className="rounded-lg border p-4">
+					<h2 className="mb-2 font-medium">API Status</h2>
+					<div className="flex items-center gap-2">
+						<div
+							className={`h-2 w-2 rounded-full ${healthCheck === "OK" ? "bg-green-500" : healthCheck === undefined ? "bg-orange-400" : "bg-red-500"}`}
+						/>
+						<span className="text-sm text-muted-foreground">
+							{healthCheck === undefined
+								? "Checking..."
+								: healthCheck === "OK"
+									? "Connected"
+									: "Error"}
+						</span>
+					</div>
+				</section>
+			</div>
+		</div>
+	);
+}
