+++ codigo-fonte/cultivo/bts.jsonc
