+++ codigo-fonte/cultivo/packages/backend/convex/healthCheck.ts
