+++ codigo-fonte/cultivo/packages/backend/convex/schema.ts
